chore: rename files to ts
