--- packages/core/src/components/TimePicker/PeriodPicker/index.js
@@ -1,7 +1,0 @@
-import { withStyles } from "@mui/styles";
-import styles from "./styles";
-import PeriodPicker from "./PeriodPicker";
-
-export default withStyles(styles, { name: "HvTimePickerPeriodPicker" })(
-  PeriodPicker
-);
