+++ packages/core/src/components/TimePicker/PeriodPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./PeriodPicker";
