+++ packages/core/src/components/TimePicker/PeriodPicker/styles.ts
