+++ packages/core/src/components/TimePicker/TimePicker.tsx
@@ -1,7 +1,6 @@
-import React, { useState, useEffect, useRef, useMemo } from "react";
+import { useState, useEffect, useRef, useMemo } from "react";
 import PropTypes from "prop-types";
 import clsx from "clsx";
-import { withStyles } from "@mui/styles";
 
 import { Time as TimeIcon } from "@hitachivantara/uikit-react-icons";
 import {
@@ -16,6 +15,8 @@ import {
   useLocale,
   HvTypography,
   useSavedState,
+  HvFormElementProps,
+  HvBaseDropdownProps,
 } from "..";
 
 import { TimePickerUnits, TimeFormat, PeriodPickerOptions } from "./enums";
@@ -27,10 +28,8 @@ import {
   getHoursForTimeFormat,
   getTimeWithFormat24,
 } from "./timePickerConverter";
-import UnitTimePicker from "./UnitTimePicker";
-import PeriodPicker from "./PeriodPicker";
-
-import styles from "./styles";
+import { UnitTimePicker } from "./UnitTimePicker";
+import { PeriodPicker } from "./PeriodPicker";
 
 const setFocusToContent = (containerRef) => {
   containerRef?.getElementsByTagName("input")[0]?.focus();
@@ -70,11 +69,127 @@ const timeIsValid = (time, timeFormat) => {
   return hourInputState && minutesInputState && secondsInputState;
 };
 
+export type HvTimePickerClassKey =
+  | "root"
+  | "input"
+  | "label"
+  | "timePopperContainer"
+  | "separator"
+  | "periodContainer"
+  | "formElementRoot"
+  | "dropdownPlaceholder"
+  | "iconBaseRoot"
+  | "error"
+  | "labelContainer"
+  | "description"
+  | "dropdownHeaderInvalid"
+  | "dropdownPlaceholderDisabled"
+  | "dropdownHeaderOpen";
+
+export interface HvTimePickerValue {
+  hours: number;
+  minutes: number;
+  seconds: number;
+  period?: "AM" | "PM";
+}
+
+export interface HvTimePickerProps
+  extends Omit<
+    HvFormElementProps,
+    "onChange" | "value" | "defaultValue" | "readOnly"
+  > {
+  /** Id to be applied to the form element root node. */
+  id?: string;
+  /** Current value of the form element. */
+  value?: HvTimePickerValue | null | undefined;
+  /** When uncontrolled, defines the initial value. */
+  defaultValue?: HvTimePickerValue | null | undefined;
+  /** Indicates that user input is required on the form element. */
+  required?: boolean;
+  /** Indicates that the form element is disabled. */
+  disabled?: boolean;
+  /** Indicates that the form element is in read only mode. */
+  readOnly?: boolean;
+  /** The placeholder value when no time is selected. */
+  placeholder?: string;
+  /** The placeholder of the hours input. */
+  hoursPlaceholder?: string;
+  /** The placeholder of the minutes input. */
+  minutesPlaceholder?: string;
+  /** The placeholder of the seconds input. */
+  secondsPlaceholder?: string;
+  /**
+   * If the time should be presented in 12 or 24 hour format.
+   * If undefined, the component will use a format according to the passed locale.
+   * If defined, it will "override" the default value given by the locale
+   */
+  timeFormat?: "12" | "24" | 12 | 24 | undefined;
+  /** Locale that will provide the time format(12 or 24 hour format). It is "overwritten" by `timeFormat` */
+  locale?: string;
+
+  /**
+   * Allow starting with an empty value by not defaulting to the current time.
+   * This should become the default behavior when the component is promoted to core.
+   */
+  disableDefaultValue?: boolean;
+
+  /**
+   * Default value for the hours picker
+   * @deprecated use defaultValue instead
+   */
+  hours?: number;
+  /**
+   * Default value for the minutes picker
+   * @deprecated use defaultValue instead
+   */
+  minutes?: number;
+  /**
+   * Default value for the seconds picker
+   * @deprecated use defaultValue instead
+   */
+  seconds?: number;
+  /**
+   * Default value for the period picker
+   * @deprecated use defaultValue instead
+   */
+  period?: "AM" | "PM";
+
+  /**
+   * Callback function to be triggered when the input value is changed.
+   * It is invoked with a object param with the following props:
+   *  - hours (in a 24h format)
+   *  - minutes
+   *  - seconds
+   *  - period
+   *
+   * It is always invoked with the hours in a 24h format
+   */
+  onChange?: (timeIn24Format: HvTimePickerValue) => void;
+
+  /**
+   * Callback called when dropdown changes the expanded state.
+   *
+   * @param {object} event The event source of the callback.
+   * @param {boolean} open If the dropdown new state is open (`true`) or closed (`false`).
+   */
+  onToggle?: (event: Event, open: boolean) => void;
+
+  /**
+   * Disable the portal behavior. The children stay within it's parent DOM hierarchy.
+   */
+  disablePortal?: boolean;
+
+  /** Sets if the calendar container should follow the date picker input out of the screen or stay visible. */
+  escapeWithReference?: boolean;
+
+  /** Extra properties to be passed to the TimePicker's dropdown. */
+  dropdownProps?: Partial<HvBaseDropdownProps>;
+}
+
 /**
  * A TimePicker component used to choose the time.
  */
-
-const HvTimePicker = ({
+export const HvTimePicker = ({
   classes,
   className,
 
@@ -123,7 +238,7 @@ const HvTimePicker = ({
   escapeWithReference = true,
   dropdownProps,
   ...others
-}) => {
+}: HvTimePickerProps) => {
   // #region STATE
   const elementId = useUniqueId(id, "hvtimepicker");
 
@@ -788,5 +903,3 @@ HvTimePicker.propTypes = {
    */
   dropdownProps: PropTypes.instanceOf(Object),
 };
-
-export default withStyles(styles, { name: "HvTimePicker" })(HvTimePicker);
