--- packages/core/src/components/TimePicker/UnitTimePicker/index.js
@@ -1,7 +1,0 @@
-import { withStyles } from "@mui/styles";
-import styles from "./styles";
-import UnitTimePicker from "./UnitTimePicker";
-
-export default withStyles(styles, { name: "HvTimePickerUnitTimePicker" })(
-  UnitTimePicker
-);
