--- packages/core/src/components/TimePicker/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from "./TimePicker";
-export * from "./TimePicker";
