--- packages/core/src/components/TimePicker/index.js
@@ -1,1 +1,0 @@
-export { default } from "./TimePicker";
